requirements.txt
requests>=2.25.1
pyyaml>=5.4.1
pytrends>=4.9.0
firecrawl-py>=1.0.0
google-generativeai>=0.8.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
